Revert "Feature/#51 review api"
