+++ src/main/java/com/weble/linkedhouse/exception/controller/ExceptionControllerAdvise.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @RestControllerAdvice
-public class ExceptionControllerAdvice {
+public class ExceptionControllerAdvise {
 
     @ExceptionHandler(LinkedHouseException.class)
     public ResponseEntity<ErrorResponse> customException(LinkedHouseException e) {
