+++ src/main/java/com/weble/linkedhouse/review/entity/FeedbackCustomer.java
@@ -1,10 +1,21 @@
-package com.weble.linkedhouse.review.domain.entity;
+package com.weble.linkedhouse.review.entity;
 
 import com.weble.linkedhouse.customer.entity.Customer;
 import com.weble.linkedhouse.house.entity.House;
 import com.weble.linkedhouse.util.AuditingFields;
-import jakarta.persistence.*;
-import lombok.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
 
 import java.util.Objects;
 
@@ -17,7 +28,7 @@ public class FeedbackCustomer extends AuditingFields {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "feedback_customer_id")
-    private Long feedbackCustomerId;
+    private Long feedbackcustomerId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "customer_id")
@@ -41,48 +52,31 @@ public class FeedbackCustomer extends AuditingFields {
     private int scoreSatisfaction;
 
     @Column(name = "total_score")
-    private double totalScore;
+    private int totalScore;
 
     @Builder
-    private FeedbackCustomer(Customer customer, House house, String title, String content,
-                             int scoreClean, int scoreCommunication, int scoreSatisfaction){
+    private FeedbackCustomer(Customer customer, House house,
+                             String title, String content, int scoreClean, int scoreCommunication
+                             , int scoreSatisfaction, int totalScore){
         this.customer = customer;
         this.house = house;
         this.title = title;
         this.content = content;
         this.scoreClean = scoreClean;
         this.scoreCommunication = scoreCommunication;
-        this.scoreSatisfaction = scoreSatisfaction;
-        this.totalScore = calculate();
+        this. scoreSatisfaction = scoreSatisfaction;
+        this.totalScore = totalScore;
     }
 
-    public static FeedbackCustomer of(Customer customer, House house, String title, String content,
-                                      int scoreClean, int scoreCommunication, int scoreSatisfaction) {
-        return new FeedbackCustomer(customer, house, title, content, scoreClean, scoreCommunication, scoreSatisfaction);
-
-    }
-
-    public void updateReview(String content, int scoreClean, int scoreCommunication, int scoreSatisfaction) {
-        this.content = content;
-        this.scoreClean = scoreClean;
-        this.scoreCommunication = scoreCommunication;
-        this.scoreSatisfaction = scoreSatisfaction;
-    }
-
-    public double calculate() {
-        return (double) (scoreSatisfaction * scoreCommunication * scoreClean) / 3.0;
-    }
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (!(o instanceof FeedbackCustomer that)) return false;
-        return this.feedbackCustomerId != null && getFeedbackCustomerId().equals(that.getFeedbackCustomerId());
+        return this.feedbackcustomerId != null && getFeedbackcustomerId().equals(that.getFeedbackcustomerId());
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(getFeedbackCustomerId());
+        return Objects.hash(getFeedbackcustomerId());
     }
-
-
 }
